Get main and reorganize changes
