+++ Assets/Prefabs/AnEnemy.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 3915b0b900a76c947b89455c201c3fe5
-folderAsset: yes
-DefaultImporter:
+guid: cacab8e47636caf4c90fc76371de6121
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
