+++ Assets/Prefabs/Enemies/Melee.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 46f503eeb1405f241acddcb1b56b4df8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
