+++ Assets/Prefabs/Enemies/Shooting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17de04e431448e3429000241ed5a5f6c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
