+++ Assets/Prefabs/Enemies/SpecialTest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6f2fcf1acb474c4aa1e3546acb24f28
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
