+++ Assets/Scripts/Enemy/EnemyCombatMelee.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f2b3c767642faf942b0bef34182f970f
+guid: d711a72b35aca09488ec33e0dbd26a56
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
