+++ Assets/Scripts/Enemy/EnemyCombatShooting.cs
@@ -2,28 +2,27 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class ShootingEnemy : MonoBehaviour
+public class EnemyCombatShooting : MonoBehaviour
 {
     [Header("Intelligence")]
     [SerializeField] private bool seeThroughWalls = false;
+    [SerializeField] private LayerMask terrainLayer;
 
     [Header("Shooting Stats")]
     [SerializeField] private int fireRate = 1;
-    [SerializeField] private float fireDelay = 0.25f;
+    [SerializeField] private float fireDelayBetweenBullet = 0.25f;
     [SerializeField] private float fireRecoveryTime;
     [SerializeField] private float fireForce;
     public EnemyProjectile projectilePrefab;
 
     private GameObject _closestPlayer;
     private bool _canFire = true;
+    private GameObject[] _players;
 
-    // Start is called before the first frame update
     void Start()
     {
-        
     }
 
-    // Update is called once per frame
     void Update()
     {
         if (_canFire)
@@ -38,9 +37,9 @@ public class ShootingEnemy : MonoBehaviour
     {
         if (!_canFire) return;
       
-        StartCoroutine(SequentialFireWithDelay(fireDelay));
+        StartCoroutine(SequentialFireWithDelay(fireDelayBetweenBullet));
 
-        // Down time before player can fire again
+        // Down time between bursts of bullets can fire again
         _canFire = false;
         StartCoroutine(RecoverFire(fireRecoveryTime));
     }
@@ -49,16 +48,18 @@ public class ShootingEnemy : MonoBehaviour
     {
         for (int i = 0; i < fireRate; i++)
         {
+            //Aim for nearest Player
             FindClosestPlayerWithTag();
 
+            //Check if Enemy can See Player
             if (HasLineOfSight() || seeThroughWalls)
             {
                 Vector2 directionToClosestPlayer = (_closestPlayer.transform.position - transform.position).normalized;
 
                 var projectile = Instantiate(projectilePrefab, transform.position, Quaternion.identity);
                 projectile.Init(directionToClosestPlayer, fireForce);
-
             }
+
             yield return new WaitForSeconds(bulletDelay);
 
         }
@@ -76,13 +77,16 @@ public class ShootingEnemy : MonoBehaviour
 
     private void FindClosestPlayerWithTag()
     {
-        GameObject[] players = GameObject.FindGameObjectsWithTag("Player");
+        _players = GameObject.FindGameObjectsWithTag("Player");
+
         float closestDistance = Mathf.Infinity;
         Vector3 currentPosition = transform.position;
 
-        foreach (GameObject player in players)
+        //For each player, check whos currently the closest
+        foreach (GameObject player in _players)
         {
             Vector3 directionToPlayer = player.transform.position - currentPosition;
+            //Gets distance of enemy to player
             float distanceToPlayer = directionToPlayer.sqrMagnitude;
 
             if (distanceToPlayer < closestDistance)
@@ -95,32 +99,27 @@ public class ShootingEnemy : MonoBehaviour
 
     private IEnumerator FindClosestPlayerCoroutine()
     {
-        yield return new WaitForSeconds(1.0f); // Adjust the interval as needed
+        yield return new WaitForSeconds(1.0f);
         FindClosestPlayerWithTag();
     }
 
     #endregion
 
     #region Line Of Sight
+
     private bool HasLineOfSight() 
     {
         Vector2 direction = transform.position - _closestPlayer.transform.position;
-        RaycastHit2D hit = Physics2D.Raycast(_closestPlayer.transform.position, direction, direction.magnitude);
-
-        // Layer mask to filter out objects you want to be considered as obstacles (e.g., walls).
-        int layerMask = LayerMask.GetMask("Terrain");
+        RaycastHit2D hit = Physics2D.Raycast(_closestPlayer.transform.position, direction, direction.magnitude, terrainLayer);
 
         // Check if the ray hits any obstacles.
-        if (hit.collider != null && (layerMask & (1 << hit.collider.gameObject.layer)) != 0)
+        if (hit.collider != null)
         {
-            // A collision occurred; there is no direct line of sight.
             return false;
         }
-        else
-        {
-            // No collision; there is a direct line of sight.
-            return true;
-        }
+        
+        return true;
     }
+
     #endregion
 }
